# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  isAdmin: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type TreeTier = typeof TreeTier[keyof typeof TreeTier];


export const TreeTier = {
  Seedling: 'Seedling',
  Classic: 'Classic',
  Premium: 'Premium',
} as const;

export type TreeStatus = typeof TreeStatus[keyof typeof TreeStatus];


export const TreeStatus = {
  available: 'available',
  adopted: 'adopted',
} as const;

export interface Tree {
  id: number;
  treeCode: string;
  name: string;
  species: string;
  tier: TreeTier;
  pricePerYear: number;
  location: string;
  estimatedHarvestKg: number;
  imageUrl: string;
  status: TreeStatus;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  co2PerYear?: number | null;
}

export type AdoptionStatus = typeof AdoptionStatus[keyof typeof AdoptionStatus];


export const AdoptionStatus = {
  pending_payment: 'pending_payment',
  active: 'active',
  expired: 'expired',
} as const;

export interface Adoption {
  id: number;
  userId: string;
  treeId: number;
  treeName: string;
  stewardName: string;
  email: string;
  phone: string;
  /** @nullable */
  dedicationMessage?: string | null;
  status: AdoptionStatus;
  /** @nullable */
  paymentId?: string | null;
  /** @nullable */
  checkoutUrl?: string | null;
  createdAt: string;
  tree?: Tree | null;
}

export interface AdoptionInput {
  treeId: number;
  /** @minLength 1 */
  stewardName: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  treeName: string;
  dedicationMessage?: string;
  isGift?: boolean;
  /** @nullable */
  recipientName?: string | null;
  /** @nullable */
  recipientEmail?: string | null;
}

export interface CheckoutInput {
  adoptionId: number;
}

export interface CheckoutSession {
  checkoutUrl: string;
  referenceNumber: string;
}

export interface JournalEntry {
  id: number;
  treeId: number;
  title: string;
  body: string;
  /** @nullable */
  photoUrl?: string | null;
  entryDate: string;
  createdAt: string;
}

export interface JournalEntryInput {
  treeId: number;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  photoUrl?: string;
  entryDate: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListTreesParams = {
type?: string;
tier?: string;
location?: string;
};