# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: Trees
    description: Tree catalog operations
  - name: Adoptions
    description: Tree adoption operations
  - name: Journal
    description: Tree farm journal entries
  - name: Admin
    description: Admin-only operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /trees:
    get:
      tags: [Trees]
      operationId: listTrees
      summary: List available trees
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: tier
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of trees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tree'
  /trees/{id}:
    get:
      tags: [Trees]
      operationId: getTree
      summary: Get a tree by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tree details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tree'
        '404':
          description: Tree not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /adoptions:
    get:
      tags: [Adoptions]
      operationId: listMyAdoptions
      summary: List adoptions for the current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of adoptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Adoption'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Adoptions]
      operationId: createAdoption
      summary: Create a new tree adoption
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdoptionInput'
      responses:
        '201':
          description: Adoption created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adoption'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /adoptions/{id}:
    get:
      tags: [Adoptions]
      operationId: getAdoption
      summary: Get an adoption by ID
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Adoption details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adoption'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /payments/checkout:
    post:
      tags: [Adoptions]
      operationId: createCheckoutSession
      summary: Create a PayMongo checkout session
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutInput'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSession'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Payment provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /payments/webhook:
    post:
      tags: [Adoptions]
      operationId: handlePaymentWebhook
      summary: Handle PayMongo webhook events
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /trees/{id}/journal:
    get:
      tags: [Journal]
      operationId: listJournalEntries
      summary: List journal entries for a tree (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JournalEntry'
        '404':
          description: Tree not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/journal:
    post:
      tags: [Admin]
      operationId: createJournalEntry
      summary: Create a journal entry (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryInput'
      responses:
        '201':
          description: Journal entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/journal/{id}:
    delete:
      tags: [Admin]
      operationId: deleteJournalEntry
      summary: Delete a journal entry (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/trees:
    get:
      tags: [Admin]
      operationId: adminListTrees
      summary: List all trees with their journal counts (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: List of trees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tree'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isAdmin]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        isAdmin:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Tree:
      type: object
      required: [id, name, species, tier, pricePerYear, location, estimatedHarvestKg, imageUrl, status, treeCode]
      properties:
        id:
          type: integer
        treeCode:
          type: string
        name:
          type: string
        species:
          type: string
        tier:
          type: string
          enum: [Seedling, Classic, Premium]
        pricePerYear:
          type: integer
        location:
          type: string
        estimatedHarvestKg:
          type: integer
        imageUrl:
          type: string
        status:
          type: string
          enum: [available, adopted]
        description:
          type: [string, 'null']
        co2PerYear:
          type: [integer, 'null']
    Adoption:
      type: object
      required: [id, userId, treeId, treeName, stewardName, email, phone, treeName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        treeId:
          type: integer
        treeName:
          type: string
        stewardName:
          type: string
        email:
          type: string
        phone:
          type: string
        dedicationMessage:
          type: [string, 'null']
        status:
          type: string
          enum: [pending_payment, active, expired]
        paymentId:
          type: [string, 'null']
        checkoutUrl:
          type: [string, 'null']
        createdAt:
          type: string
        tree:
          oneOf:
            - $ref: '#/components/schemas/Tree'
            - type: 'null'
    AdoptionInput:
      type: object
      required: [treeId, stewardName, email, phone, treeName]
      properties:
        treeId:
          type: integer
        stewardName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
          minLength: 1
        treeName:
          type: string
          minLength: 1
        dedicationMessage:
          type: string
        isGift:
          type: boolean
        recipientName:
          type: [string, 'null']
        recipientEmail:
          type: [string, 'null']
    CheckoutInput:
      type: object
      required: [adoptionId]
      properties:
        adoptionId:
          type: integer
    CheckoutSession:
      type: object
      required: [checkoutUrl, referenceNumber]
      properties:
        checkoutUrl:
          type: string
        referenceNumber:
          type: string
    JournalEntry:
      type: object
      required: [id, treeId, title, body, entryDate, createdAt]
      properties:
        id:
          type: integer
        treeId:
          type: integer
        title:
          type: string
        body:
          type: string
        photoUrl:
          type: [string, 'null']
        entryDate:
          type: string
          format: date
        createdAt:
          type: string
    JournalEntryInput:
      type: object
      required: [treeId, title, body, entryDate]
      properties:
        treeId:
          type: integer
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        photoUrl:
          type: string
        entryDate:
          type: string
          format: date