# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  treeId: number;
  title: string;
  body: string;
  /** @nullable */
  photoUrl?: string | null;
  entryDate: Date;
  createdAt: string;
}