# lib/api-zod/src/generated/types/journalEntryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntryInput {
  treeId: number;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  photoUrl?: string;
  entryDate: Date;
}